requirements.txt
torch==2.4.1
torchvision==0.19.1
tensorflow<=2.13.1
stardist==0.8.5
csbdeep==0.8.0
numpy==1.24.3
umap-learn==0.5.5
hdbscan==0.8.38.post1
pandas==2.0.3
tika==2.6.0
opencv-python==4.10.0.84
wandb==0.17.8
seaborn==0.13.2
alphashape==1.3.1
skops==0.9.0
natsort==8.4.0
alive-progress==3.1.5
reportlab==4.2.2
openpyxl==3.1.5
timm==0.5.4
pydicom==2.4.4
matplotlib==3.7.5
numba==0.58.1
pillow==10.4.0
pyyaml==6.0.1
scikit-learn==1.3.2
tabulate==0.9.0
tqdm==4.66.5
-e git+https://github.com/IPMI-ICNS-UKE/sparsam.git@f43c6e645194aa2030a837621f6a598806004aeb#egg=sparsam






